requirements.txt
flask
flask-cors
cloudinary
python-dotenv
gunicorn
